Reviews Added
Added a section within the details page to show reviews published by users of the application
