+++ app/build.gradle.kts
@@ -69,6 +69,7 @@ android {
 dependencies {
     implementation("androidx.core:core-ktx:1.9.0")
 //    implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.6.1")
+    implementation("androidx.compose.material3:material3:1.0.0-alpha07")
     implementation("androidx.activity:activity-compose:1.7.0")
     implementation(platform("androidx.compose:compose-bom:2023.03.00"))
     implementation("androidx.core:core-ktx:1.12.0")
