+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.SpotifyClone"
         tools:targetApi="31">
         <activity
-            android:name="com.example.spotifyclone.MainActivity"
+            android:name="com.example.bookcraftapplication.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.SpotifyClone">
