+++ app/src/main/java/com/example/bookcraftapplication/auth/ResultAuth.kt
@@ -1,0 +1,6 @@
+package com.example.bookcraftapplication.auth
+
+sealed class ResultAuth<out T> {
+    data class Success<out T>(val data: T) : ResultAuth<T>()
+    data class Failure(val exception: Throwable) : ResultAuth<Nothing>()
+}
