+++ app/src/main/java/com/example/bookcraftapplication/auth/User.kt
@@ -1,0 +1,3 @@
+package com.example.bookcraftapplication.auth
+
+data class User(var email: String)
