+++ app/src/main/java/com/example/bookcraftapplication/data/ProfileData.kt
@@ -1,0 +1,6 @@
+package com.example.bookcraftapplication.data
+
+data class ProfileData  (
+    var name: String = "",
+    var counter: Int = 0
+)
