--- app/src/main/java/com/example/spotifyclone/auth/ResultAuth.kt
@@ -1,6 +1,0 @@
-package com.example.spotifyclone.auth
-
-sealed class ResultAuth<out T> {
-    data class Success<out T>(val data: T) : ResultAuth<T>()
-    data class Failure(val exception: Throwable) : ResultAuth<Nothing>()
-}
