--- app/src/main/java/com/example/spotifyclone/auth/User.kt
@@ -1,3 +1,0 @@
-package com.example.spotifyclone.auth
-
-data class User(var email: String)
