--- app/src/main/java/com/example/spotifyclone/data/ProfileData.kt
@@ -1,6 +1,0 @@
-package com.example.spotifyclone.data
-
-data class ProfileData  (
-    var name: String = "",
-    var counter: Int = 0
-)
