--- app/src/main/java/com/example/spotifyclone/data/ProfileRepository.kt
@@ -1,9 +1,0 @@
-package com.example.spotifyclone.data
-
-import kotlinx.coroutines.flow.Flow
-
-interface ProfileRepository {
-    suspend fun saveProfile(profileData: ProfileData)
-    fun getProfile(): Flow<ProfileData>
-    suspend fun clear()
-}
