--- app/src/main/java/com/example/spotifyclone/ui/theme/Color.kt
@@ -1,74 +1,0 @@
-package com.example.spotifyclone.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-val earthy_brown = Color(0xFF8B4513)
-val forest_green = Color( 0xFF8F9779)
-val md_theme_light_onPrimary = Color(0xFFFFFFFF)
-val md_theme_light_primaryContainer = Color(0xFF89F8C7)
-val md_theme_light_onPrimaryContainer = Color(0xFF002114)
-val md_theme_light_secondary = Color(0xFF4D6357)
-val md_theme_light_onSecondary = Color(0xFFFFFFFF)
-val md_theme_light_secondaryContainer = Color(0xFFCFE9D9)
-val md_theme_light_onSecondaryContainer = Color(0xFF092016)
-val md_theme_light_tertiary = Color(0xFF3D6373)
-val md_theme_light_onTertiary = Color(0xFFFFFFFF)
-val md_theme_light_tertiaryContainer = Color(0xFFC1E8FB)
-val md_theme_light_onTertiaryContainer = Color(0xFF001F29)
-val md_theme_light_error = Color(0xFFBA1A1A)
-val md_theme_light_errorContainer = Color(0xFFFFDAD6)
-val md_theme_light_onError = Color(0xFFFFFFFF)
-val md_theme_light_onErrorContainer = Color(0xFF410002)
-val md_theme_light_background = Color(0xFFFBFDF9)
-val md_theme_light_onBackground = Color(0xFF191C1A)
-val md_theme_light_surface = Color(0xFFFBFDF9)
-val md_theme_light_onSurface = Color(0xFF191C1A)
-val md_theme_light_surfaceVariant = Color(0xFFDBE5DD)
-val md_theme_light_onSurfaceVariant = Color(0xFF404943)
-val md_theme_light_outline = Color(0xFF707973)
-val md_theme_light_inverseOnSurface = Color(0xFFEFF1ED)
-val md_theme_light_inverseSurface = Color(0xFF2E312F)
-val md_theme_light_inversePrimary = Color(0xFF6CDBAC)
-val md_theme_light_shadow = Color(0xFF000000)
-val md_theme_light_surfaceTint = Color(0xFF006C4C)
-val md_theme_light_outlineVariant = Color(0xFFBFC9C2)
-val md_theme_light_scrim = Color(0xFF000000)
-
-val md_theme_dark_primary = Color(0xFF6CDBAC)
-val md_theme_dark_onPrimary = Color(0xFF003826)
-val md_theme_dark_primaryContainer = Color(0xFF005138)
-val md_theme_dark_onPrimaryContainer = Color(0xFF89F8C7)
-val md_theme_dark_secondary = Color(0xFFB3CCBE)
-val md_theme_dark_onSecondary = Color(0xFF1F352A)
-val md_theme_dark_secondaryContainer = Color(0xFF354B40)
-val md_theme_dark_onSecondaryContainer = Color(0xFFCFE9D9)
-val md_theme_dark_tertiary = Color(0xFFA5CCDF)
-val md_theme_dark_onTertiary = Color(0xFF073543)
-val md_theme_dark_tertiaryContainer = Color(0xFF244C5B)
-val md_theme_dark_onTertiaryContainer = Color(0xFFC1E8FB)
-val md_theme_dark_error = Color(0xFFFFB4AB)
-val md_theme_dark_errorContainer = Color(0xFF93000A)
-val md_theme_dark_onError = Color(0xFF690005)
-val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
-val md_theme_dark_background = Color(0xFF191C1A)
-val md_theme_dark_onBackground = Color(0xFFE1E3DF)
-val md_theme_dark_surface = Color(0xFF191C1A)
-val md_theme_dark_onSurface = Color(0xFFE1E3DF)
-val md_theme_dark_surfaceVariant = Color(0xFF404943)
-val md_theme_dark_onSurfaceVariant = Color(0xFFBFC9C2)
-val md_theme_dark_outline = Color(0xFF8A938C)
-val md_theme_dark_inverseOnSurface = Color(0xFF191C1A)
-val md_theme_dark_inverseSurface = Color(0xFFE1E3DF)
-val md_theme_dark_inversePrimary = Color(0xFF006C4C)
-val md_theme_dark_shadow = Color(0xFF000000)
-val md_theme_dark_surfaceTint = Color(0xFF6CDBAC)
-val md_theme_dark_outlineVariant = Color(0xFF404943)
-val md_theme_dark_scrim = Color(0xFF000000)
